extracted profile dropdown to component
